Improves general error detection and reporting
We have extended the try-catch block in order to capture errors when
including *Rest files, which may connect to the database, and thus,
if the database is not available an uncaught exception is thrown.
We now also capture these kind of exceptions in our catch block.

We have also improved the error reporting by sending a JSON object
with an "error" property including the error message.